+++ ExilenceNextApp/src/components/table-wrapper/TableWrapper.styles.ts
@@ -101,38 +101,3 @@ export const useStyles = makeStyles((theme: Theme) =>
     },
   })
 );
-
-const areEqual = (prevProps: any, nextProps: any) =>
-  prevProps.checked === nextProps.checked && prevProps.indeterminate === nextProps.indeterminate;
-
-export const HeaderCheckbox = React.memo(
-  styled(Checkbox)({
-    fontSize: '1rem',
-    margin: '-8px 0 -8px -15px',
-    padding: '8px 9px',
-    '& svg': {
-      width: '24px',
-      height: '24px',
-    },
-    '&:hover': {
-      backgroundColor: 'transparent',
-    },
-  }),
-  areEqual
-);
-
-export const RowCheckbox = React.memo(
-  styled(Checkbox)({
-    fontSize: '14px',
-    margin: '-9px 0 -8px -15px',
-    padding: '8px 9px 9px 9px',
-    '&:hover': {
-      backgroundColor: 'transparent',
-    },
-    '& svg': {
-      width: 24,
-      height: 24,
-    },
-  }),
-  areEqual
-);
